fix style corp
