+++ solutions/Virtual-Assistant/src/csharp/skills/tests/emailskilltest/API/Fakes/MSGraph/MockGraphServiceClient.cs
