+++ solutions/Virtual-Assistant/src/csharp/skills/tests/emailskilltest/Flow/Utterances/BaseTestUtterances.cs
@@ -43,8 +43,10 @@ namespace EmailSkillTest.Flow.Utterances
 
         public Email GetBaseNoneIntent()
         {
-            var emailIntent = new Email();
-            emailIntent.Intents = new Dictionary<Intent, IntentScore>();
+            var emailIntent = new Email
+            {
+                Intents = new Dictionary<Intent, IntentScore>()
+            };
             emailIntent.Intents.Add(Intent.None, new IntentScore() { Score = TopIntentScore });
 
             return emailIntent;
@@ -61,22 +63,25 @@ namespace EmailSkillTest.Flow.Utterances
             string[] subject = null,
             string[] message = null)
         {
-            var emailIntent = new Email();
+            var emailIntent = new Email
+            {
+                Text = userInput,
 
-            emailIntent.Text = userInput;
-
-            emailIntent.Intents = new Dictionary<Intent, IntentScore>();
+                Intents = new Dictionary<Intent, IntentScore>()
+            };
             emailIntent.Intents.Add(intent, new IntentScore() { Score = TopIntentScore });
 
-            emailIntent.Entities = new _Entities();
-            emailIntent.Entities._instance = new _Entities._Instance();
+            emailIntent.Entities = new _Entities
+            {
+                _instance = new _Entities._Instance(),
 
-            emailIntent.Entities.ordinal = ordinal;
-            emailIntent.Entities.number = number;
-            emailIntent.Entities.ContactName = contactName;
-            emailIntent.Entities.SenderName = senderName;
-            emailIntent.Entities.EmailSubject = subject;
-            emailIntent.Entities.Message = message;
+                ordinal = ordinal,
+                number = number,
+                ContactName = contactName,
+                SenderName = senderName,
+                EmailSubject = subject,
+                Message = message
+            };
 
             if (!string.IsNullOrEmpty(userInput))
             {
